Servo code added and commented.
git-svn-id: https://cc3.svn.sourceforge.net/svnroot/cc3/trunk@77 111db651-790d-0410-a044-b0abe44a5820

+++ hal/lpc2106-cmucam3/servo.h
@@ -6,15 +6,27 @@
 #include "inttypes.h"
 #include "stdbool.h"
 
+// SERVO_RESOLUTION sets resolution within 1ms
+// Failure may occure above 512
 #define SERVO_RESOLUTION 255 
 #define MAX_SERVOS 4
 #define SERVO_PERIOD 18
 
-// SERVO_RESOLUTION sets resolution within 1ms
+// Sets up the servo timers and begins servicing the servos
 void cc3_servo_init ();
+
+// Interrupt called from interrupt.c when the servo needs to be serviced
 void _cc3_servo_int ();
-void _cc3_servo_lo (int n);
+
+// Set a particular servo pin low
+void _cc3_servo_lo (uint8_t n);
+
+// Set all pins high at the start of the servo cycle
 void _cc3_servo_hi_all ();
-int cc3_servo_set (int servo, int pos);
+
+// User function to set a servo
+uint8_t cc3_servo_set (uint8_t servo, uint32_t pos);
+
+// User function to disable servos to conserve power
 int cc3_servo_disable ();
 #endif
